documentation fixes
More information needs to be added to the "Features" section. It is important for users to see all the features so they know what they're getting and can use the application to its fullest potential. Help commands should be listed as well as any other features missing in the list and instructions on how to read the visual graphs should be put in somewhere.